Refactored poin in polygon analysis
